feat: featured and new templates
