+++ apps/journeys-admin/src/components/TemplateSections/FeaturedAndNewTemplates/index.ts
@@ -1,0 +1,1 @@
+export { FeaturedAndNewTemplates } from './FeaturedAndNewTemplates'
